Data count for barcounts

This function is intended to work in combination with geom_bar to display the sum of the values represented by each bar. Like geom_bar, this function works if the x and/or y aesthetics are provided.

geom_barcount(
  mapping = NULL,
  data = NULL,
  stat = "barcount",
  position = "stack",
  ...,
  overall.stack = TRUE,
  digits = 3,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_barcount(
  mapping = NULL,
  data = NULL,
  geom = "barcount",
  position = "stack",
  ...,
  overall.stack = TRUE,
  digits = 3,
  width = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping

Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

position

Position adjustment, either as a string, or the result of a call to a position adjustment function.

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

overall.stack

Defines whether an overall count is displayed for stacked bars or if the counts of each individual component of the stacked bars should be displayed.

digits

Integer indicating the number of significant digits to be used. Recognized values are 0..22. Use digits = 0 to display as integers.

width

Bar width. By default, set to 90% of the resolution of the data.

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

orientation

The orientation of the layer. The default (NA) automatically determines the orientation from the aesthetic mapping. In the rare event that this fails it can be given explicitly by setting orientation to either "x" or "y". See the Orientation section for more detail.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

geom, stat

Use to override the default connection between geom_barcount and stat_barcount.

Orientation

This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the orientation parameter, which can be either "x" or "y". The value gives the axis that the geom should run along, "x" being the default orientation you would expect for the geom.

See also

Examples

if (FALSE) {
p <- ggplot(mpg)
p +
  aes(x = class) +
  geom_bar() +
  geom_barcount()

# Map class to y instead to flip the orientation
p +
  aes(y = class) +
  geom_bar() +
  geom_barcount()

# For stacked position
p <- ggplot(diamonds, aes(color, fill = cut))
p +
  geom_bar(position = 'stack') +
  geom_barcount()
p +
  geom_bar(position = 'stack') +
  geom_barcount(overall.stack = FALSE)

# For dodged position
p +
  geom_bar(position = 'dodge') +
  geom_barcount(position = position_dodge(width = 0.9))

# For fill position
p +
  geom_bar(position = 'fill') +
  geom_barcount(position = position_fill())

# For fillpercent position
p +
  geom_bar(position = 'fillpercent') +
  geom_barcount(position = position_fillpercent()) +
  ylab('count (%)')

}