Data count for histograms

This function is intended to work in combination with geom_histogram to display the sum of the values represented by each bar.

The same non-default arguments used in the the geom_histogram call should also be used in the geom_histcount call.

Density/percentage can be displayed by setting the y aesthetic to after_stat(density / after_stat(percent) (which are provided by the bin2 stat) and the label aesthetic to after_stat(density_label) / after_stat(percent_label) (which are provided by the histcount stat).

geom_histcount(
  mapping = NULL,
  data = NULL,
  stat = "histcount",
  position = "stack",
  ...,
  digits = 3,
  binwidth = NULL,
  bins = NULL,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_histcount(
  mapping = NULL,
  data = NULL,
  geom = "histcount",
  position = "stack",
  ...,
  digits = 3,
  binwidth = NULL,
  bins = NULL,
  center = NULL,
  boundary = NULL,
  breaks = NULL,
  closed = c("right", "left"),
  pad = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping

Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

position

Position adjustment, either as a string, or the result of a call to a position adjustment function.

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

digits

Integer indicating the number of significant digits to be used. Recognized values are 0..22. Use digits = 0 to display as integers.

binwidth

The width of the bins. Can be specified as a numeric value or as a function that calculates width from unscaled x. Here, "unscaled x" refers to the original x values in the data, before application of any scale transformation. When specifying a function along with a grouping structure, the function will be called once per group. The default is to use the number of bins in bins, covering the range of the data. You should always override this value, exploring multiple widths to find the best to illustrate the stories in your data.

The bin width of a date variable is the number of days in each time; the bin width of a time variable is the number of seconds.

bins

Number of bins. Overridden by binwidth. Defaults to 30.

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

orientation

The orientation of the layer. The default (NA) automatically determines the orientation from the aesthetic mapping. In the rare event that this fails it can be given explicitly by setting orientation to either "x" or "y". See the Orientation section for more detail.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

geom, stat

Use to override the default connection between geom_histcount and stat_histcount.

center

bin position specifiers. Only one, center or boundary, may be specified for a single plot. center specifies the center of one of the bins. boundary specifies the boundary between two bins. Note that if either is above or below the range of the data, things will be shifted by the appropriate integer multiple of binwidth. For example, to center on integers use binwidth = 1 and center = 0, even if 0 is outside the range of the data. Alternatively, this same alignment can be specified with binwidth = 1 and boundary = 0.5, even if 0.5 is outside the range of the data.

boundary

bin position specifiers. Only one, center or boundary, may be specified for a single plot. center specifies the center of one of the bins. boundary specifies the boundary between two bins. Note that if either is above or below the range of the data, things will be shifted by the appropriate integer multiple of binwidth. For example, to center on integers use binwidth = 1 and center = 0, even if 0 is outside the range of the data. Alternatively, this same alignment can be specified with binwidth = 1 and boundary = 0.5, even if 0.5 is outside the range of the data.

breaks

Alternatively, you can supply a numeric vector giving the bin boundaries. Overrides binwidth, bins, center, and boundary.

closed

One of "right" or "left" indicating whether right or left edges of bins are included in the bin.

pad

If TRUE, adds empty bins at either end of x. This ensures frequency polygons touch 0. Defaults to FALSE.

Orientation

This geom treats each axis differently and, thus, can thus have two orientations. Often the orientation is easy to deduce from a combination of the given mappings and the types of positional scales in use. Thus, ggplot2 will by default try to guess which orientation the layer should have. Under rare circumstances, the orientation is ambiguous and guessing may fail. In that case the orientation can be specified directly using the orientation parameter, which can be either "x" or "y". The value gives the axis that the geom should run along, "x" being the default orientation you would expect for the geom.

Aesthetics

geom_histcount() understands the following aesthetics (required aesthetics are in bold):

x

y

label

alpha

angle

colour

family

fontface

group

hjust

lineheight

size

vjust

Learn more about setting these aesthetics in vignette("ggplot2-specs").

See also

Examples

if (FALSE) {
p <- ggplot(diamonds)

# Histogram for continuous variable count
p +
  aes(x = price) +
  geom_histogram() +
  geom_histcount()

# Map class to y instead to flip the orientation
p +
  aes(y = price) +
  geom_histogram() +
  geom_histcount()

# Histogram with a fill aesthetic
p +
  aes(x = price, fill = clarity) +
  geom_histogram() +
  geom_histcount()

# Histogram for continuous variable density
p +
  aes(x = price) +
  geom_histogram(aes(y = after_stat(density)), stat = 'bin2', bins = 15) +
  geom_histcount(aes(y = after_stat(density), label = after_stat(density_label)), bins = 15)

# Histogram for continuous variable percentage using the bin2 stat
p +
  aes(x = price) +
  geom_histogram(aes(y = after_stat(percent)), stat = 'bin2', bins = 15) +
  geom_histcount(aes(y = after_stat(percent), label = after_stat(percent_label)), bins = 15) +
  ylab('percent (%)')
}